# config/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Watson wherespace",
        "description": "",
        "version": "1.0.0",
        "contact":{
        "name": "OpenNTF",
        "url": "http://www.openntf.org",
        "email": "info@openntf.org"
        }
    },
    "host": "api.rabobank.n",
    "schemes": [
        "https"
    ],
    "basePath": "/v2",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/person/{personId}": {
            "get": {
                "summary": "Get the person by his id",
                "description": "Return the Event by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "The ID of the person.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Person"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid person ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create the person by his id",
                "description": "Create the Person by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "The ID of the person.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Person"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid person ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the person by his id",
                "description": "Update the Person by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "The ID of the person.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Person"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid person ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete the person by his id",
                "description": "Delete the Person by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "The ID of the person.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Person"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "400": {
                        "description": "Invalid person ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/person/email":{
            "get":{
                "summary": "Get events by person",
                "description": "Get the person by email",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "email",
                        "description": "Fetched all the person by email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Person"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid email supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Person": {
            "type": "object",
            "properties": {
                "personId": {
                    "type": "string"
                },
                "emailaddress": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "office": {
                    "type": "string"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        },
        "Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}